+++ digits/changelog.md
@@ -2,6 +2,8 @@
 *Non-trivial pull requests should include an entry below. Entries must be suitable for inclusion in public-facing materials such as release notes and blog posts. Keep them short, sweet, and in the past tense. New entries go on top. When merging to deploy, add the version number and date.*
 
 ## Unreleased
+* (IC) [Fixed crash on StateButton when multiple clicks] (https://github
+.com/twitter/digits-android/issues/2)
 
 ## v1.4.2
 * (IC) Added normalized phone number from /device/login endpoint